requirement.txt
# ETL Pipeline Dependencies
pandas>=2.0.0
luigi>=3.4.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0